Refactored test
