+++ lib/utilities.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+// ignore: avoid_classes_with_only_static_members
 /// This class contains general static utility methods used throughout the
 /// app.
 class Utilities {
